requirements.txt
Django>=5.0,<6.0
django-ninja>=1.4.0
pydantic>=2.0

locust>=2.0

djangorestframework-simplejwt>=5.0

psycopg2-binary>=2.9

python-dotenv>=1.0.0
